requirements.txt
# Demo MCP System - Dependencies

# Core Framework
streamlit>=1.28.0
fastmcp>=2.0.0
httpx>=0.25.0
pyyaml>=6.0.1

# Data Processing
numpy>=1.24.0
pydantic>=2.5.0

# Optional: Azure OpenAI (for production use)
# openai>=1.0.0
# azure-identity>=1.15.0

# Optional: Vector Database (for production use)
# faiss-cpu>=1.7.4
# qdrant-client>=1.7.0

# Testing (development)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Note: The system works in demo mode without Azure OpenAI
# FastMCP handles OpenAPI spec loading and tool exposure via MCP protocol